refactor: 구분자 관련 기능 적절한 클래스로 이동
