+++ src/test/java/calculator/controller/StringCalculatorControllerTest.java
@@ -1,8 +1,5 @@
 package calculator.controller;
 
-import calculator.model.CustomDelimiter;
-import calculator.model.DefaultDelimiter;
-import calculator.model.Delimiter;
 import calculator.model.InputData;
 import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.BeforeEach;
@@ -29,22 +26,6 @@ class StringCalculatorControllerTest extends NsTest {
             assertThat(stringCalculatorController.handoverInput())
                     .isInstanceOf(InputData.class);
         });
-    }
-
-    @Test
-    @DisplayName("커스텀 구분자가 있는 데이터 객체를 받아 커스텀 구분자 객체를 생성한다")
-    void createCustomDelimiterInstance() {
-        InputData inputData = new InputData("//;\\n1;2;3");
-        assertThat(stringCalculatorController.createDelimiterPart(inputData))
-                .isInstanceOf(CustomDelimiter.class);
-    }
-
-    @Test
-    @DisplayName("커스텀 구분자가 없는 데이터 객체를 받아 기본 구분자 객체를 생성한다")
-    void createDefaultDelimiterInstance() {
-        InputData inputData = new InputData("1,2,3");
-        assertThat(stringCalculatorController.createDelimiterPart(inputData))
-                .isInstanceOf(DefaultDelimiter.class);
     }
 
     @Override
